Refactor GameEvent to use ints; add ApplyMove to GameEngine
Refactored GameEvent to use int properties instead of strings for player, edgeId, boxX, boxY, and nextPlayer, improving type safety. Updated all related unit tests accordingly. Enhanced GameEngine by updating SetEdgeOwner and GetEdgeOwner to use int, initializing edge owner arrays with Board12x12 constants, and adding a new ApplyMove method for move validation and event tracking. Introduced a gameEvents list to record move history and added a Reset method stub. Included minor code cleanups and helper method usage.

+++ DotsAndBoxes.Game.Core.Test/GameEngineTests.cs
@@ -30,7 +30,7 @@ namespace DotsAndBoxes.Game.Core.Test
             // Arrange
             var gameEngine = new GameEngine();
             int edgeId = 0;
-            sbyte player = 0;
+            int player = 0;
 
             // Act
             gameEngine.SetEdgeOwner(
